requirements/dev.txt
alembic==1.4.1
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.1
click==7.1
coverage==5.0.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
regex==2020.2.20
six==1.14.0
SQLAlchemy==1.3.13
toml==0.10.0
typed-ast==1.4.1
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.1.0
celery==4.4.1